move importmap to html reference analysis
